requirements.txt
# InboxHunter Agent - Dependencies
# ================================

# Browser Automation
playwright==1.56.0
playwright-stealth==2.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.0
aiofiles==23.2.1

# WebSocket
websockets==12.0

# System Tray UI
pystray==0.19.5
Pillow==10.2.0

# Configuration & Validation
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
pyyaml==6.0.1

# CAPTCHA Services
2captcha-python==1.2.1

# LLM Providers
openai>=1.0.0

# Utilities
python-dateutil==2.8.2
phonenumbers==8.13.29

# Logging
loguru==0.7.2
colorama==0.4.6

# Build Tools
pyinstaller==6.3.0

# Optional: For development
# -------------------------
# pytest==7.0.0
# pytest-asyncio==0.23.4
# black==24.1.1
# flake8==7.0.0
